requirements.txt
APScheduler
black
datasets
gradio
gradio[oauth]
gradio_leaderboard==0.0.13
gradio_client
huggingface-hub>=0.18.0
matplotlib
numpy
pandas
python-dateutil
tqdm
transformers
tokenizers>=0.15.0
sentencepiece
litellm==1.60.8
inspect_ai==0.3.63
pydantic==2.10.6